requirements.txt
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
colorama==0.4.3
cryptography==3.1.1
decorator==4.4.2
dnspython==1.16.0
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
PySocks==1.7.1
python-telegram-bot==12.8
six==1.15.0
toml==0.10.1
tornado==6.0.4
wrapt==1.12.1
